// packages/web/src/api/schema.d.ts
export interface components {
	schemas: {
// ...
		Conclusion: {
			/** Id */
			id: string;
			/** Content */
			content: string;
			/**
			 * Observer Id
			 * @description The peer who made the conclusion
			 */
			observer_id: string;
			/**
			 * Observed Id
			 * @description The peer the conclusion is about
			 */
			observed_id: string;
			/** Session Id */
			session_id?: string | null;
			/**
			 * Created At
			 * Format: date-time
			 */
			created_at: string;
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}